Connect db models to app
App routes now query db
